back/python/django_server/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
uWSGI==2.0.21
